fix: circular dep in eth service
